# Change time format to unix time.
# POST /api/v1.0/users/:user-id/asks
Add a field common, which is the channel name. While it will not be used in this version we will need it soon.

# GET /api/v1.0/users/:user-id/feed
## Add a field number of replies - put a dummy value 1 or two randomly
## Add a sub document for the users who replied.
## Add a field common, which is the name of the channel. Default value is friends
## Add a field in the user object - relationship. It should get three value: Friend, Friend of Friend and Other.
Default is Other

* GET /api/v1.0/users/:user-id/asks/:ask-status
Add a field common, which is the name of the channel. Default value is friends
Add a field in the user object - relationship. It shoudl get three value: Friend, Friend of Friend and Other.
Default is Other